ReportTrainee refactored to extend from new Report class, which extends Page. ReportTrainer created and also extends Report
